Create greeting GUI program
