--- prac_07/box_layout_demo.py
@@ -1,12 +1,0 @@
-from kivy.app import App
-from kivy.lang import Builder
-
-
-class BoxLayoutDemo(App):
-    def build(self):
-        self.title = "Box Layout Demo"
-        self.root = Builder.load_file('box_layout.kv')
-        return self.root
-
-
-BoxLayoutDemo().run()
